requirements.txt
bitsandbytes==0.42.0
decord==0.6.0
deepspeed==0.14.4
einops==0.8.0
einops_exts==0.0.4
flash_attn==2.6.1
huggingface_hub==0.23.4
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
packaging==24.2
peft==0.11.1
Pillow==11.0.0
pyav==13.1.0
qwen_vl_utils==0.0.8
Requests==2.32.3
ring_flash_attn==0.1.1
tokenizers
tqdm==4.66.4
transformers
xformers==0.0.28.post3